refactor(users): extract magic-login consumption into _magic_login.py (Issue 649)
Move the magic-login redemption machinery — _current_jwt_user,
_validate_magic_user, _consume_magic_token, and the GET
/magic-login/{token}/ route — out of _auth.py into a dedicated
_magic_login.py module with its own router, keeping _auth.py under the
500-line limit. Pure restructuring: endpoint path, rate limit, auth, and
responses are unchanged.

_auth.py drops from 497 to 372 lines; the new router is wired into
users/api.py alongside the existing sub-routers.
