+++ backend/users/api.py
@@ -9,6 +9,7 @@ from users._helpers import (  # re-exported
     _validate_member_role_required,
 )
 from users._magic_links import router as magic_links_router
+from users._magic_login import router as magic_login_router
 from users._management import router as management_router
 from users._roles import router as roles_router
 
@@ -22,5 +23,6 @@ __all__ = [
 router = Router()
 router.add_router("", roles_router)
 router.add_router("", auth_router)
+router.add_router("", magic_login_router)
 router.add_router("", management_router)
 router.add_router("", magic_links_router)
